mypy

+++ main.py
@@ -1,7 +1,7 @@
 import math
 
 
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> int:
     return math.floor(a + b)
